requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and Vector Store
qdrant-client==1.6.9
# sqlite3 is built-in, no need to install

# Document Processing
unstructured[all-docs]==0.11.2
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.1.0

# Text Processing and NLP
langchain==0.0.340
langchain-community==0.0.1
sentence-transformers==2.2.2
jieba==0.42.1
rank-bm25==0.2.2
spacy==3.7.2

# AI API Integration
openai==1.3.7

# Caching and Storage
redis==5.0.1
aiofiles==23.2.1

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1